+++ src/Pages/Games.tsx
@@ -3,8 +3,8 @@ import Col from "react-bootstrap/Col";
 import { ListItem } from "../partials/ListItem";
 import { Link } from "react-router-dom";
 import { useI18n } from "../providers/I18nProvider";
-import * as en from "../texts/games/en.json";
-import * as de from "../texts/games/de.json";
+import * as en from "../assets/texts/games/en.json";
+import * as de from "../assets/texts/games/de.json";
 import robotsImgUrl from "/src/assets/img/robots_env.png";
 import wickedtdImgUrl from "/src/assets/img/wickedtd_cropped.png";
 import breadImgUrl from "/src/assets/img/Villa_Final.jpg";
