+++ src/assets/texts/games/de.json
